+++ jni/java/com/wallet/crypto/trustapp/jni/P2SHPrefix.java
@@ -8,13 +8,13 @@ package com.wallet.crypto.trustapp.jni;
 
 
 public enum P2SHPrefix {
-    BITCOIN (0x05),
-    LITECOIN (0x32),
-    DASH (0x10);
+    BITCOIN ((byte) 0x05),
+    LITECOIN ((byte) 0x32),
+    DASH ((byte) 0x10);
 
-    private final int value;
-    P2SHPrefix(int value) {
+    private final byte value;
+    P2SHPrefix(byte value) {
         this.value = value;
     }
-    public int value() { return value; }
+    public byte value() { return value; }
 }
