+++ swift/Sources/P2SHPrefix.swift
@@ -4,7 +4,7 @@
 // terms governing use, modification, and redistribution, is contained in the
 // file LICENSE at the root of the source code distribution tree.
 
-public enum P2SHPrefix: Int {
+public enum P2SHPrefix: UInt8 {
     case bitcoin = 0x05
     case litecoin = 0x32
     case dash = 0x10
